+++ dEventer/app/src/main/java/es/achraf/deventer/view/adapters/AdapterMensajes.java
@@ -1,4 +1,4 @@
-package es.achraf.deventer.adaptadores;
+package es.achraf.deventer.view.adapters;
 
 import android.annotation.SuppressLint;
 import android.app.Notification;
